Move in-memory internals to dedicated namespace
Refactored several internal classes and records from M.EventBrokerSlim.Internal to M.EventBrokerSlim.Internal.InMemory. This improves code organization by separating in-memory implementations from other internal components. No functional changes were made; only namespace declarations were updated.
