+++ EventBrokerSlim/src/M.EventBrokerSlim/Internal/InMemory/DynamicHandlerClaimTicket.cs
@@ -1,5 +1,5 @@
 ﻿using System;
 
-namespace M.EventBrokerSlim.Internal;
+namespace M.EventBrokerSlim.Internal.InMemory;
 
 internal sealed record DynamicHandlerClaimTicket(Guid Id, Type EventType) : IDynamicHandlerClaimTicket;
