+++ EventBrokerSlim/src/M.EventBrokerSlim/Internal/InMemory/RetryQueue.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 
-namespace M.EventBrokerSlim.Internal;
+namespace M.EventBrokerSlim.Internal.InMemory;
 
 internal class RetryQueue
 {
